+++ src/main/java/ar/edu/unlam/tallerweb1/servicios/ServicioTurnos.java
@@ -9,10 +9,10 @@ public interface ServicioTurnos {
 	
 	List<Medico> consultarMedico();
 	
-	List <Medico> listaDeMedicosPorEspecialidad (Integer especialidadId);
+	List <Medico> listaDeMedicosPorEspecialidad (Long especialidadId);
 
 	List<String> turnosDeMedicoEspecifico(Medico medico);
 
-	List<String> listaDeDiasDisponibles(Integer especialidadId);
+	List<String> listaDeDiasDisponibles(Long especialidadId);
 
 }
